refactor: update Board and Console components to use selectors and clean up unused code
